Kód obsluhující data přihlášeného uživatele přesunut do nové třídy User.
